Refactor controllers and views
https://trello.com/c/reemF8gj/159-checks-on-identity-questions-in-tool
