+++ app/views/assessments/activity/activity-1.html.erb
