+++ app/views/assessments/activity/activity-2.html.erb
