+++ app/views/assessments/activity/activity-3.html.erb
