+++ app/views/assessments/activity/activity-result.html.erb
