+++ app/views/assessments/activity/activity-start.html.erb
